+++ Inc/i2sportadapter.hpp
@@ -91,7 +91,7 @@ class I2sPortAdapter : public AudioPortAdapterStrategy {
 
     /**
      * @brief Return the size of the one sample in I2S frame.
-     * @return 2, 3 or 4. The unit is [Byte]
+     * @return 1..32. The unit is [bit]
      */
     virtual unsigned int GetSampleDataSizeTx();
     /**
@@ -102,7 +102,7 @@ class I2sPortAdapter : public AudioPortAdapterStrategy {
 
     /**
      * @brief Return the size of the one sample in I2S frame.
-     * @return 2, 3 or 4. The unit is [Byte]
+     * @return 1..32. The unit is [bit]
      */
     virtual unsigned int GetSampleDataSizeRx();
     /**